:recycle: refactor!: reorganize the version structure
