+++ lib/changelog/generate.js
@@ -2,9 +2,9 @@ import fs from "node:fs";
 import path from "node:path";
 
 import { categorizeCommits, normalizeCommits } from "../commits.js";
-import { computeVersion } from "../compute-version.js";
 import { config } from "../config.js";
 import { getLastTag, getCommits, parseCommit } from "../git.js";
+import { computeVersion } from "../version/compute.js";
 import {
   hasAnyContent,
   insertReleaseSection,
