+++ lib/dry-run.js
@@ -2,9 +2,9 @@ import fs from "node:fs";
 import path from "node:path";
 
 import { generateChangelog } from "./changelog/generate.js";
-import { computeVersion } from "./compute-version.js";
 import { config } from "./config.js";
 import { generateReleaseNotes } from "./release-notes.js";
+import { computeVersion } from "./version/compute.js";
 
 /* ===========================
  * Core API (Programmatic)
