+++ lib/release-notes.js
@@ -1,7 +1,6 @@
 import fs from "node:fs";
 import path from "node:path";
 
-import { computeVersion } from "./compute-version.js";
 import { config } from "./config.js";
 import { getLastTag } from "./git.js";
 import {
@@ -10,6 +9,7 @@ import {
   getDefaultBranch,
   generateReleaseNotesViaAPI,
 } from "./github.js";
+import { computeVersion } from "./version/compute.js";
 
 /* ===========================
  * Core API (Programmatic)
