+++ lib/version/compute.js
@@ -1,7 +1,7 @@
-import { config } from "./config.js";
-import { getLastTag, getCommits, parseCommit } from "./git.js";
-import { readPackageVersion } from "./utils.js";
-import { bumpVersion, detectBumpType } from "./versioning.js";
+import { config } from "../config.js";
+import { getLastTag, getCommits, parseCommit } from "../git.js";
+import { readPackageVersion } from "../utils.js";
+import { bumpVersion, detectBumpType } from "../versioning.js";
 
 /* ===========================
  * Core API (Programmatic)
